requirements.txt
Django
psycopg2
django-widget-tweaks
django-allauth
Click
pillow
sqlalchemy
django[argon2]